Multi-fix for netware.
Now allows empty import/export files to be handled which then terminate the current directive processing properly rather than reading through into non-related tokens.

Fixes issues with the symbol prefixing used in Novell. The prefix was held in a global so a nested file would inherit the current prefix of its parent file if set. This is obviously not what was intended. The prefix is now attached to the current command file and is reset in each call to ProcNovImport and ProcNovExport. This allows for nested files to define their own prefixes and not interfere with / be interfered with by other files.


git-svn-id: http://svn.cmeerw.net/openwatcom@2992 9848d260-d8c2-4c77-ac60-93a84d57cbad