ChangeLog: The CauseWay DLL support has been updated. Several problems have been corrected and the user-supplied startup/termination function is now called DllMain() rather than main(); the latter had undesirable side effects.
Implementation note: The assembly startup code was horrendously buggy.
The code was restructured to more resemble OS/2 and NT startup, with
most of the termination and some of the initialization code moved to C
modules. The dllstart module might potentially be made part of the
standard clib, obviating the need for explicit linking.


git-svn-id: http://svn.cmeerw.net/openwatcom@12293 9848d260-d8c2-4c77-ac60-93a84d57cbad